saved sticking page naive implementation
