+++ src/components/UserDashboard/SavedStickings/StickingsRow.tsx
@@ -2,13 +2,17 @@ import { FaRegTrashAlt } from 'react-icons/fa';
 import { FaPlay } from 'react-icons/fa6';
 import { Divider } from '@mui/material';
 import Button from '../../Button/Button';
+import { useNavigate } from 'react-router';
 interface Props {
   sticking: { [key: string]: string };
   onDelete: (id: string) => void;
 }
 const StickingsRow = ({ sticking, onDelete }: Props) => {
+  const navigate = useNavigate();
+
   const stickingValue: { [key: string]: string } = Object(sticking.sticking);
   const stickingText = `${stickingValue['beat-1']} - ${stickingValue['beat-2']} - ${stickingValue['beat-3']} - ${stickingValue['beat-4']} `;
+
   return (
     <>
       <li style={{ padding: '0.5rem' }}>
@@ -16,7 +20,13 @@ const StickingsRow = ({ sticking, onDelete }: Props) => {
           <span style={{ flex: '1', padding: '0.5rem 0rem' }}>
             {stickingText}
           </span>
-          <Button idName="play-sticking-btn" children={<FaPlay />} />
+          <Button
+            idName="play-sticking-btn"
+            children={<FaPlay />}
+            onBtnClick={() =>
+              navigate('/user-account/single-sticking', { replace: true })
+            }
+          />
           <Button
             idName="delete-sticking-btn"
             children={<FaRegTrashAlt />}
