+++ src/pages/SingleSticking/SingleStickings.css
@@ -1,4 +1,4 @@
-/* .container {
+.container {
   display: flex;
   flex-direction: column;
   justify-content: space-around;
@@ -46,4 +46,4 @@
     height: 80vh;
     justify-content: flex-start;
   }
-} */
+}
